backend/refactor: backend/refactor: dashboard driver module decomposition
